Finished API Build

+++ src/main/java/com/ecommerce/repository/UserRepository.java
@@ -1,8 +1,17 @@
 package com.ecommerce.repository;
 
-import org.springframework.data.repository.CrudRepository;
+import java.util.List;
 
+import org.springframework.data.jdbc.repository.query.Query;
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.repository.query.Param;
+
+import com.ecommerce.model.Product;
+import com.ecommerce.model.Purchase;
 import com.ecommerce.model.User;
 
-public interface UserRepository extends CrudRepository<User, Integer> {}
+public interface UserRepository extends CrudRepository<User, Long> {	
+	
+	
+}
